affe/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "affe"
version = "0.1.0"
description = "AFFE manager agent with Outlook subagent"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "msal>=1.24.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "flask>=2.0.0",
]

[project.scripts]
affe = "affe.main:main"
affe-ui = "affe.ui:main"

[tool.setuptools.packages.find]
where = ["src"]
